refactor(webhooks): move WebhookError to dedicated error.rs
